requirements.txt
discord.py>=2.4.0
websockets>=12.0
httpx>=0.27.0
fastapi==0.116.1
google-genai>=1.56.0
langgraph==1.0.3
openai==1.97.0
anthropic>=0.42.0
pydantic==2.11.7
pydantic-settings==2.5.2
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
uvicorn==0.35.0

# SAIMemory embeddings
fastembed>=0.7.3
numpy>=1.26.0

# Image processing
Pillow>=11.0.0
markdownify>=0.12.0

# llama.cpp integration for local GGUF models
llama-cpp-python>=0.3.0

# Linting
ruff>=0.9.0
